+++ server/src/main/java/org/allaymc/server/registry/loader/WorldStorageFactoryRegistryLoader.java
@@ -3,6 +3,7 @@ package org.allaymc.server.registry.loader;
 import org.allaymc.api.registry.RegistryLoader;
 import org.allaymc.api.world.storage.WorldStorage;
 import org.allaymc.server.world.storage.AllayLevelDBWorldStorage;
+import org.allaymc.server.world.storage.AllayMemoryWorldStorage;
 
 import java.nio.file.Path;
 import java.util.HashMap;
@@ -17,6 +18,7 @@ public class WorldStorageFactoryRegistryLoader implements RegistryLoader<Void, M
     public Map<String, Function<Path, WorldStorage>> load(Void $) {
         var map = new HashMap<String, Function<Path, WorldStorage>>();
         map.put("LEVELDB", AllayLevelDBWorldStorage::new);
+        map.put("MEMORY", AllayMemoryWorldStorage::new);
         return map;
     }
 }