+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Unless otherwise specified, any version comparison below is the comparison of se
 
 <small>[Compare with 0.1.3](https://github.com/AllayMC/Allay/compare/0.1.3...HEAD)</small>
 
+### Added
+
+- Support storage world in memory by using `storage-type:MEMORY`.
+
 ## [0.1.3](https://github.com/AllayMC/Allay/releases/tag/0.1.3) (API 0.4.0) - 2025-1-17
 
 <small>[Compare with 0.1.2](https://github.com/AllayMC/Allay/compare/0.1.2...0.1.3)</small>